+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Stop rendering widgets when `BBox` is empty
 - Made sure to pass out `BBox` to `Renderer::render_done` method